@@ -1,8 +1,14 @@
 module github.com/libdns/duckdns
 
-go 1.14
+go 1.18
 
 require (
-	github.com/libdns/libdns v0.2.0
+	github.com/libdns/libdns v1.0.0-beta.1
 	github.com/miekg/dns v1.1.40
+)
+
+require (
+	golang.org/x/crypto v0.0.0-20191011191535-87dc89f01550 // indirect
+	golang.org/x/net v0.0.0-20190923162816-aa69164e4478 // indirect
+	golang.org/x/sys v0.0.0-20190924154521-2837fb4f24fe // indirect
 )
